lite --> nodom, pull out to shared file
